# openapi/go-api.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "\nTraceRTM is a comprehensive requirements traceability management system that helps teams\ntrack relationships between requirements, features, and tasks throughout the development lifecycle.\n\n## Features\n- Project and Item Management\n- Link Management for traceability\n- AI Agent Coordination\n- Graph-based relationship analysis\n- Full-text search with pgvector\n- Real-time updates via WebSocket and NATS\n\n## Authentication\nCurrently supports WorkOS AuthKit. Set Authorization header with Bearer token.",
    "title": "TraceRTM API",
    "contact": {
      "name": "TraceRTM Support",
      "url": "https://github.com/kooshapari/tracertm-backend",
      "email": "support@tracertm.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/api/v1/docs": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Lists all documentation for a project with pagination",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "documentation"
        ],
        "summary": "List documentation",
        "parameters": [
          {
            "type": "string",
            "description": "Project ID",
            "name": "project_id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "Limit (default 100)",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Offset (default 0)",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of documentation",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to list documentation",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Parses and indexes documentation in various formats (markdown, rst, etc.)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "documentation"
        ],
        "summary": "Index new documentation",
        "parameters": [
          {
            "description": "Documentation to index",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handlers.IndexDocumentationRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Documentation indexed successfully",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to index documentation",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      }
    },
    "/api/v1/docs/search": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Performs full-text search on documentation",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "documentation"
        ],
        "summary": "Search documentation",
        "parameters": [
          {
            "type": "string",
            "description": "Project ID",
            "name": "project_id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Search query",
            "name": "query",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "Limit (default 100)",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Offset (default 0)",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to search documentation",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      }
    },
    "/api/v1/docs/{id}": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Retrieves a specific documentation by ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "documentation"
        ],
        "summary": "Get documentation by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Documentation ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Documentation retrieved",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "404": {
            "description": "Documentation not found",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to get documentation",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Updates an existing documentation",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "documentation"
        ],
        "summary": "Update documentation",
        "parameters": [
          {
            "type": "string",
            "description": "Documentation ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated documentation",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handlers.IndexDocumentationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Documentation updated",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "404": {
            "description": "Documentation not found",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to update documentation",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Deletes a documentation record (soft delete)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "documentation"
        ],
        "summary": "Delete documentation",
        "parameters": [
          {
            "type": "string",
            "description": "Documentation ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Documentation deleted",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "404": {
            "description": "Documentation not found",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to delete documentation",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "handlers.ErrorResponse": {
      "description": "Error response with error message",
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
...
    "handlers.IndexDocumentationRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "file_path": {
          "type": "string"
        },
        "format": {
          "description": "markdown, rst, html, plaintext",
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": true
        },
        "project_id": {
          "type": "string"
        },
        "source_url": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "description": "Enter your bearer token in the format: Bearer {token}",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}